requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
werkzeug
cryptography
pytz